Playlists the user doesn't own are now correctly filtered out.
The cause of this error was due to the fact I was trying to delete elements
from a list in a loop. Everytime an element was deleted, the list was
shifted. Very hard to diagnose, very easy to fix. It simply involved
parsing through a copy of the list, rather than the original contents. Very
happy to have finally fixed this bug after a few months.